fix: Move ready queue processing after identity request
